ASR/requirements.txt
sounddevice==0.4.6
numpy>=1.21.0
librosa>=0.9.2
webrtcvad>=2.0.10
torch>=1.10.0
torchaudio>=0.10.0
funasr>=0.6.0
pyyaml>=6.0 